апввапва

+++ field.py
@@ -1,15 +1,15 @@
 import numpy as np
 from models import GameState, Point3D, Snake, Food, Enemy
 
-array = np.array((1, 1, 1), dtype=np.int8)
+array = np.array((1, 1, 1), dtype=np.int32)
 
 def create_field(map_size_x, map_size_y, map_size_z):
     global array
-    array = np.ones((map_size_x, map_size_y, map_size_z), dtype=np.int8)
+    array = np.ones((map_size_x, map_size_y, map_size_z), dtype=np.int32)
 
 def poke_anything(list_cord: Point3D, value: int):
     global array
-    array[list_cord.root[0]][list_cord.root[1]][list_cord.root[2]] = np.int8(value)
+    array[list_cord.root[0]][list_cord.root[1]][list_cord.root[2]] = np.int32(value)
 
 def poke_all_oranges(oranges: list[Food]):
     for orange in oranges:
